Create README.md
The sailor coin originates from the revival vision of global nautical culture and is positioned as a bridge connecting the marine economy and the digital world. It is widely used in shipping services, marine conservation projects, shipping finance and other fields, and realizes transaction transparency and security through blockchain technology innovation. The community ecology is booming, bringing together nautical enthusiasts and investors from all over the world. Join hands with well-known shipping enterprises and ecological and environmental protection organizations as partners to jointly promote the sustainable development of the ocean. The seaman token aims to create a prosperous and green marine digital economy future.